Update ndvi_anomaly to use dask
It looks like the underlying process has changed significantly since the
original script was written. This commit updates it as well as
converting it to use dask.

Two new params added: crs and mosaic_type.

This commit also adds a `mask_good_quality` function. This is used in
all of the notebooks so it's pulled into a separate file.