Align data exports with source notebooks
Change geotiff creation so it aligns properly with the data exports from
the original notebooks.

Most of the data we're exporting is already a DataSet so we don't want
to try to convert it to one. Most of the time when the data is a
DataArray we're happy letting `export_xarray_to_geotiff` convert the
DataArray directly to a geotiff without first converting it to a DataSet
ourselves.